Issue #6728 - QUIC and HTTP/3
- Implemented support for GREASE for unidirectional and bidirectional streams.
- Implemented UnknownBodyParser.
- Strengthened validation of frames in control and message streams.

+++ jetty-http3/http3-common/src/main/java/org/eclipse/jetty/http3/internal/DecoderStreamConnection.java
@@ -24,7 +24,7 @@ import org.eclipse.jetty.io.EndPoint;
 public class DecoderStreamConnection extends InstructionStreamConnection
 {
     // SPEC: QPACK Encoder Stream Type.
-    public static final int STREAM_TYPE = 0x03;
+    public static final long STREAM_TYPE = 0x03;
 
     private final QpackEncoder encoder;
