+++ jetty-http3/http3-common/src/main/java/org/eclipse/jetty/http3/internal/EncoderStreamConnection.java
@@ -24,7 +24,7 @@ import org.eclipse.jetty.io.EndPoint;
 public class EncoderStreamConnection extends InstructionStreamConnection
 {
     // SPEC: QPACK Encoder Stream Type.
-    public static final int STREAM_TYPE = 0x02;
+    public static final long STREAM_TYPE = 0x02;
 
     private final QpackDecoder decoder;
